Fix FOUC of dark mode
For astro I had to change the script to is:inline to stop it from
being preprocessed and not running in time.
But then I couldn't change the innerHtml content of the button, which
pretty janky to begin with.
So I moved the svg to be a background image so that it is no longer
dependent on JS, which was too slow.